unify habit modals (create/update) - also reorganized all habit widgets into subdir
