+++ lib/widgets/habit/habit_card.dart
@@ -6,7 +6,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_font_icons/flutter_font_icons.dart';
 import 'package:vibration/vibration.dart';
 
-import 'edit_habit_modal.dart';
+import 'habit_modal.dart';
 
 class HabitCard extends StatelessWidget {
   final Habit habit;
@@ -31,7 +31,7 @@ class HabitCard extends StatelessWidget {
         showModalBottomSheet(
           backgroundColor: Colors.transparent,
           context: context,
-          builder: (context) => EditHabitModal(habit),
+          builder: (context) => HabitModal(habit: habit),
         );
       },
       child: Container(
